pyproject.toml
[project]
name = "chat-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
]
